moved step on and two into one visual tile. updated functionality to work appropriately
